+++ bigtree/__init__.py
@@ -1,7 +1,7 @@
 __version__ = "0.3.2"
 
 from bigtree.dag.construct import dataframe_to_dag, dict_to_dag, list_to_dag
-from bigtree.dag.export import dag_to_dot
+from bigtree.dag.export import dag_to_dataframe, dag_to_dict, dag_to_dot, dag_to_list
 from bigtree.node.basenode import BaseNode
 from bigtree.node.dagnode import DAGNode
 from bigtree.node.node import Node