+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tree Exporter: Support Nodes with same name.
 - Tree Modifier: Modify docstring.
 - Utility Iterators: Modify docstring.
+- DAG Exporter: To list, nested dictionary, pandas DataFrame.
 
 ## [0.3.2] - 2022-11-07
 ### Added
@@ -32,7 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Node: DAGNode for creating DAG graph.
 - Tree Exporter: Support Node to dot which can export to png, jpeg, dot files.
-- Tree Exporter: Support DAGNode to dot.
+- DAG Exporter: Support DAGNode to dot.
 
 ## [0.1.0] - 2022-11-01
 ### Added